requirements.txt
groq>=0.9.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
gitpython>=3.1.43
python-dotenv>=1.0.1
